style(#280): 欠点豆図鑑のフィルター・検索・ソートを統合モーダルに集約してヘッダー競合を解消
